+++ ColorfulPieces/ComfyLib/Commands/ComfyCommand.cs
@@ -11,7 +11,7 @@ public sealed class ComfyCommand : Attribute {
 }
 
 public static class ComfyCommandUtils {
-  static readonly List<Terminal.ConsoleCommand> _commands = new();
+  static readonly List<Terminal.ConsoleCommand> _commands = [];
 
   public static void ToggleCommands(bool toggleOn) {
     DeregisterCommands(_commands);
@@ -42,8 +42,7 @@ public static class ComfyCommandUtils {
     return assembly.GetTypes()
         .SelectMany(type => type.GetMethods(BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic))
         .Where(method => method.GetCustomAttributes(typeof(ComfyCommand), inherit: false).Length > 0)
-        .Where(method => IsRegisterCommandMethod(method) || IsRegisterCommandsMethod(method))
-        .SelectMany(method => RegisterCommands(method));
+        .SelectMany(RegisterCommands);
   }
 
   static IEnumerable<Terminal.ConsoleCommand> RegisterCommands(MethodInfo method) {