+++ ColorfulPieces/Patches/HudPatch.cs
@@ -10,24 +10,23 @@ using static PluginConfig;
 
 [HarmonyPatch(typeof(Hud))]
 static class HudPatch {
-  static readonly string HoverNameTextTemplate =
+  public static readonly string HoverNameTextTemplate =
     "{0}{1}"
-        + "<size={9}>"
-        + "[<color={2}>{3}</color>] Set piece color: <color=#{4}>#{4}</color> (<color=#{4}>{5}</color>)\n"
-        + "[<color={6}>{7}</color>] Clear piece color\n"
-        + "[<color={6}>{8}</color>] Copy piece color\n"
+        + "<size={8}>"
+        + "[<color={2}>{3}</color>] Change color: <color=#{4}>#{4}</color> (<color=#{4}>{5}</color>)\n"
+        + "[<color={6}>{7}</color>] Clear color\n"
         + "</size>";
 
   [HarmonyPostfix]
   [HarmonyPatch(nameof(Hud.UpdateCrosshair))]
-  static void UpdateCrosshairPostfix(ref Hud __instance, ref Player player) {
-    if (!IsModEnabled.Value || !ShowChangeRemoveColorPrompt.Value || !Player.m_localPlayer.Ref()?.m_hovering) {
-      return;
-    }
-
-    WearNTear wearNTear = player.m_hovering.GetComponentInParent<WearNTear>();
-
-    if (!wearNTear.Ref()?.m_nview || !wearNTear.m_nview.IsValid()) {
+  static void UpdateCrosshairPostfix(Hud __instance, Player player) {
+    if (!IsModEnabled.Value
+        || !ShowChangeRemoveColorPrompt.Value
+        || !player.m_hovering
+        || !player.m_hovering.TryGetComponentInParent(out WearNTear wearNTear)
+        || !wearNTear
+        || !wearNTear.m_nview
+        || !wearNTear.m_nview.IsValid()) {
       return;
     }
 
@@ -42,7 +41,6 @@ static class HudPatch {
             TargetPieceEmissionColorFactor.Value.ToString("N2"),
             "#EF5350",
             ClearPieceColorShortcut.Value,
-            CopyPieceColorShortcut.Value,
             ColorPromptFontSize.Value);
   }
 }