+++ ColorfulPieces/Patches/PlayerPatch.cs
@@ -15,11 +15,11 @@ static class PlayerPatch {
   [HarmonyPatch(nameof(Player.Update))]
   static IEnumerable<CodeInstruction> UpdateTranspiler(IEnumerable<CodeInstruction> instructions) {
     return new CodeMatcher(instructions)
-        .MatchForward(
-            useEnd: false,
-            new CodeMatch(OpCodes.Ldarg_0),
+        .Start()
+        .MatchStartForward(
             new CodeMatch(OpCodes.Call, AccessTools.Method(typeof(Player), nameof(Player.UpdateHover))))
-        .Advance(offset: 2)
+        .ThrowIfInvalid($"Could not patch Player.Update()! (UpdateHover)")
+        .Advance(offset: 1)
         .InsertAndAdvance(
             new CodeInstruction(OpCodes.Ldloc_1),
             Transpilers.EmitDelegate(UpdateHoverPostDelegate))